// csp-app/src/app/models/checklist-execution.ts
import { ObservationModel, AuditSampleModel } from "./audit-checklist-based-model";
import { EmpInfoModel } from "./emp-info-model";

export class CheckListExecutionModel {
    id: number;
    customeR_ID: string;
    projecT_ID: string;
    planneD_AUDIT_START_DATE: string;
    planneD_AUDIT_END_DATE: string;
    actuaL_AUDIT_START_DATE: string;
    actuaL_AUDIT_END_DATE: string;
    audiT_PLANNED_HOURS: number;
    audiT_ACTUAL_HOURS: number;
    audiT_TITLE: string;
    versionid: number;
    auditoR_NAME: number;
    audiT_SCOPE: string;
    auditeE_NAME: number[];
    cC_EMP_LIST: number[];
    tO_EMP_LIST : number[];
    applicablE_QUESTIONS: number;
    status: string;
    notes: string;
    findingS_ID: number;
    createD_DATE: Date;
    createD_BY: string;
    updateD_DATE: Date;
    updateD_BY: string;
    procesS_MODEL_DESCRIPTION: string;
    procesS_AREA_DESCRIPTION: string;
    procesS_DESCRIPTION: string;
    statuS_CATEGORY: string;
    score : number;
    maX_SCORE : number;
    weightagE_ID: number;
    weightagE_TITLE: string;
    looK_FOR: string;
    weightagE_SCORE : number;
    // findingS_DESCRIPTION : string;
    // findingS_TYPE : string;
    findings: ObservationModel[] = [];
    checklisT_STATUS_LIST_VALUES: any[];
    checklisT_STATUS_LIST_ID: number;
    issubmitted: boolean;
    servicE_AREA_ID: number;
    procesS_ID: number;
    procesS_MODEL_ID : number;
    checklisT_SAMPLE_AUDITED: AuditSampleModel[] = [];
    mappeD_CHECKLIST : boolean;
    currenT_STATUS : string;
}